+++ HISTORY.md
@@ -27,6 +27,7 @@
 * morphic: tweaked slider transparency for flat design mode
 * objects: translation support for category labels in the unified palette
 * threads: fixed vanishing JSF block bug in new error messages
+* gui: fixed blocks cache invalidation for unified palette
 
 ### 2021-07-08
 * objects: added category labels to unified palette